+++ web/vistas/vistasPersona/listar.jsp
@@ -15,6 +15,27 @@
         <link href="css/bootstrap.css" rel="stylesheet" type="text/css"/>
         <script src="js/BuscadorTabla.js" type="text/javascript"></script>
         <title>JSP Page</title>
+         <script type="text/javascript">
+            function confirmarEliminar()
+            {
+                var respuesta = confirm("¿Estás seguro que desees eliminar el registro?");
+                if (respuesta == true) {
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+
+            function confirmarEdicion()
+            {
+                var respuesta = confirm("¿Estás seguro que desees editar el registro?");
+                if (respuesta == true) {
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+        </script>
     </head>
     <body>
         <div class="container">
@@ -69,8 +90,8 @@
                         <td class="text-center"><%= per.getEmail()%></td>
                         <td class="text-center"><%= per.getIdUsuario()%></td>
                         <td class="text-center">
-                            <a class="btn btn-warning" href="ControladorPersona?accion=editar&idPersona=<%= per.getIdPersona()%>">Editar</a>
-                            <a class="btn btn-danger" href="ControladorPersona?accion=eliminar&idPersona=<%= per.getIdPersona()%>">Remove</a>
+                            <a class="btn btn-warning" href="ControladorPersona?accion=editar&idPersona=<%= per.getIdPersona()%>" onclick="return confirmarEdicion()">Editar</a>
+                            <a class="btn btn-danger" href="ControladorPersona?accion=eliminar&idPersona=<%= per.getIdPersona()%>" onclick="return confirmarEliminar()">Remove</a>
                         </td>
                     </tr>
                     <%}%>
